+++ packages/docs-data/navHelpers.test.ts
@@ -18,6 +18,8 @@ import { readFileSync } from "node:fs";
 
 import { describe, expect, it } from "@blueprintjs/test-commons/vitest";
 
+import { slugify } from "@blueprintjs/docs-theme";
+
 import {
     assignRoutes,
     buildNavLeafPage,
@@ -29,7 +31,6 @@ import {
     getSectionRefs,
     normalizeNavConfig,
     requirePage,
-    slugify,
 } from "./navHelpers.mts";
 import { PACKAGES, SECTIONS } from "./navTypes.mts";
 import type {
